refactor: split monolithic user and activity services into granular, organized service objects
