+++ app/api/activity.py
@@ -2,34 +2,38 @@ from fastapi import APIRouter, Depends, Header, Path
 from typing import List, Annotated
 
 from app.models.activity import ActivityCreate, ActivityRead
-from app.services.activity import ActivityService
+from app.services.activities.create import Create
+from app.services.activities.find_by_user import FindByUser
+from app.services.activities.find_by_user_and_program import FindByUserAndProgram
 
 router = APIRouter(tags=["Activity"])
 
-ActivityServiceDep = Annotated[ActivityService, Depends()]
+CreateServiceDep = Annotated[Create, Depends()]
+FindByUserServiceDep = Annotated[FindByUser, Depends()]
+FindByUserAndProgramServiceDep = Annotated[FindByUserAndProgram, Depends()]
 
 @router.get("/activities", response_model=List[ActivityRead])
 async def get_activities_by_user(
-    service: ActivityServiceDep,
+    service: FindByUserServiceDep,
     x_slack_user_id: str = Header(...),
 ):
-    return await service.find_activities_by_user(slack_id=x_slack_user_id)
+    return await service.execute(slack_id=x_slack_user_id)
 
 
 @router.get("/programs/{program_id}/activities", response_model=List[ActivityRead])
 async def get_activities_by_user_and_program(
-    service: ActivityServiceDep,
+    service: FindByUserAndProgramServiceDep,
     program_id: int = Path(..., title="ID Program", ge=1),
     x_slack_user_id: str = Header(...),
 ):
-    return await service.find_activities_by_user_and_program(program_id, x_slack_user_id)
+    return await service.execute(program_id, x_slack_user_id)
 
 
 @router.post("/programs/{program_id}/activities", response_model=ActivityRead)
 async def create_activity(
-    service: ActivityServiceDep,
+    service: CreateServiceDep,
     activity_create: ActivityCreate,
     program_id: int = Path(..., title="ID Program", ge=1),
     x_slack_user_id: str = Header(...),
 ):
-    return await service.insert_activity(activity_create, program_id, x_slack_user_id)
+    return await service.execute(activity_create, program_id, x_slack_user_id)
