+++ app/api/user.py
@@ -2,17 +2,19 @@ from fastapi import APIRouter, Depends
 from typing import List, Annotated
 
 from app.models.user import UserCreate, UserRead
-from app.services.user import UserService
+from app.services.users.find_all import FindAll
+from app.services.users.create import Create
 
 router = APIRouter(tags=["User"])
 
-UserServiceDep = Annotated[UserService, Depends()]
+FindAllServiceDep = Annotated[FindAll, Depends()]
+CreateServiceDep = Annotated[Create, Depends()]
 
 @router.get("/users", response_model=List[UserRead])
-async def get_users(service: UserServiceDep):
-    return await service.find_all_users()
+async def get_users(service: FindAllServiceDep):
+    return await service.execute()
 
 
 @router.post("/users", response_model=UserRead)
-async def create_user(user: UserCreate, service: UserServiceDep):
-    return await service.insert_new_user(user)
+async def create_user(user: UserCreate, service: CreateServiceDep):
+    return await service.execute(user)
